Add support for ALTER INDEX SET TABLESPACE
Summary:
Introduces support for `ALTER INDEX SET TABLESPACE`, which changes the tablespace an index is assigned to. As
with `ALTER TABLE SET TABLESPACE`, the command is asynchronous--it returns immediately, and tablets are moved by
the load balancer in the background if necessary.

A beta feature warning appears when using this command. This can be suppressed by setting the
`ysql_beta_feature_tablespace_alteration` flag to true.

This diff also adds tests for `ALTER INDEX SET TABLESPACE` and `ALTER MATERIALIZED VIEW SET TABLESPACE`.

Test Plan:
- Added PG Regress tests to `yb_tablespaces` for `ALTER INDEX SET TABLESPACE`
- Added PG Regress tests to `yb_matview` for `ALTER MATERIALIZED VIEW SET TABLESPACE`
- Added Java tests to `TestTablespaceProperties`
  - `testAlterTableSetTablespace`: Verifies tablet movement after `ALTER TABLE SET TABLESPACE`
  - `testAlterIndexSetTablespace`: Verifies tablet movement after `ALTER INDEX SET TABLESPACE`
  - `testAlterMatViewSetTablespace`: Verifies tablet movement after `ALTER MATERIALIZED VIEW SET TABLESPACE`
  - `testTablespaceLeaderPreference`: Checks that leader preferences are followed after `ALTER TABLE SET
TABLESPACE`
  - `testTablespacePartitioning`: Tests `ALTER TABLE/INDEX SET TABLESPACE` on partitioned table/indexes
  - `testAlterTableSetTablespaceConcurrent`: Tests `ALTER TABLE/INDEX/MATERIALIZED VIEW SET TABLESPACE` running
concurrently with a DML workload.

Differential Revision: https://phorge.dev.yugabyte.com/D28128